fix: 배터리 전류 소수 정밀도 보존

+++ android/app/src/main/java/com/han/battery/data/model/BatteryTelemetryBatchPayload.kt
@@ -13,7 +13,7 @@ data class BatteryTelemetryLog(
     val timestamp: Long,
     val level: Int,
     val voltage: Int,
-    val current: Int,
+    val current: Double,
     val temperature: Int,
     @SerializedName("elapsed_ms") val elapsedMs: Long,
     @SerializedName("is_charging") val isCharging: Boolean, // ⭐ 백엔드와 맞춘 스네이크 케이스
